+++ packages/wallet-connectors/src/BrowserWallet.ts
@@ -1,5 +1,10 @@
-import { SchemaType, WalletApi, detectConcordiumProvider } from '@concordium/browser-wallet-api-helpers';
-import { AccountTransactionPayload, AccountTransactionSignature, AccountTransactionType } from '@concordium/web-sdk';
+import {
+    SchemaType,
+    SendTransactionPayload,
+    WalletApi,
+    detectConcordiumProvider,
+} from '@concordium/browser-wallet-api-helpers';
+import { AccountTransactionSignature, AccountTransactionType } from '@concordium/web-sdk';
 import {
     SignableMessage,
     TypedSmartContractParameters,
@@ -87,24 +92,20 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
     }
 
     /**
-     * Returns a gRPC client that is ready to perform requests against some Concordium Node connected to network/chain
-     * that the connected account lives on.
+     * Returns the transport object of the gRPC client that the Browser Wallet uses to perform requests
+     * against some Concordium Node connected to network/chain that the connected account lives on.
      * The client implements version 2 of the Node API.
      *
      * This method is included because it's part of the Browser Wallet API.
      * It should be used with care as it's hard to guarantee that it actually connects to the expected network.
-     * The recommended alternative is to have the application instantiate its own instance
-     * that is independent of any connection.
-     * See {@link Network.grpcOpts} for more details.
-     *
-     * Implementation detail: The method cannot be moved to {@link BrowserWalletConnector}
-     * as the Browser Wallet's RPC client doesn't work until a connection has been established.
+     * The recommended alternative is to construct your own client using {@link Network.grpcOpts} which is
+     * independent of any connection.
      *
      * @return The Browser Wallet's internal gRPC client.
      * @throws If the installed version of the Browser Wallet doesn't support the method.
      */
-    getGrpcClient() {
-        return this.client.getGrpcClient();
+    getGrpcTransport() {
+        return this.client.grpcTransport;
     }
 
     /**
@@ -126,7 +127,7 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
     async signAndSendTransaction(
         accountAddress: string,
         type: AccountTransactionType,
-        payload: AccountTransactionPayload,
+        payload: SendTransactionPayload,
         typedParams?: TypedSmartContractParameters
     ): Promise<string> {
         if ((type === AccountTransactionType.InitContract || type === AccountTransactionType.Update) && typedParams) {
@@ -135,8 +136,8 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
                 case 'ModuleSchema':
                     return this.client.sendTransaction(
                         accountAddress,
-                        type,
-                        payload,
+                        type as any, // wallet API types enforce strict coupling of transaction types and corresponding payloads.
+                        payload as any, // wallet API types enforce strict coupling of transaction types and corresponding payloads.
                         parameters,
                         {
                             type: SchemaType.Module,
@@ -145,7 +146,7 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
                         schema.version
                     );
                 case 'TypeSchema':
-                    return this.client.sendTransaction(accountAddress, type, payload, parameters, {
+                    return this.client.sendTransaction(accountAddress, type as any, payload as any, parameters, {
                         type: SchemaType.Parameter,
                         value: schema.value.toString('base64'),
                     });
@@ -156,7 +157,7 @@ export class BrowserWalletConnector implements WalletConnector, WalletConnection
         if (typedParams) {
             throw new Error(`'typedParams' must not be provided for transaction of type '${type}'`);
         }
-        return this.client.sendTransaction(accountAddress, type, payload);
+        return this.client.sendTransaction(accountAddress, type as any, payload as any); // wallet API types enforce strict coupling of transaction types and corresponding payloads.
     }
 
     async signMessage(accountAddress: string, msg: SignableMessage): Promise<AccountTransactionSignature> {