+++ packages/react-components/CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
 
 ### Changed
 
 -   Build as ES module to facilitate tree-shaking.
+-   (breaking) Make compatible with `@concordium/web-sdk@^7`, which is now a peer dependency.
+-   `useContractSelector` (breaking): Support for supplying `JsonRpcClient` has been removed.
 
 ### Fixed